+++ src/app/update/page.tsx
@@ -1,1 +1,3 @@
-
+export default function UpdatePage() {
+  return
+}
